changed while condition

+++ 0x13-bit_manipulation/5-flip_bits.c
@@ -16,7 +16,7 @@ unsigned int flip_bits(unsigned long int n, unsigned long int m)
 	unsigned int counter = 0;/*This is for our loop*/
 	unsigned long int flipper = n ^ m;/*num you get after n^m*/
 
-	while (counter < sizeof(unsigned int) * 8)
+	while (counter < sizeof(unsigned long int) * 8)
 	{
 		if ((flipper & 1) == 1)/*num compares and == 1*/
 			numofbits++; /*add 1 to numofbits flipped*/
